refactor(inspections-api): extract hierarchy sub-router
Units tree, observer links, and report versions (11 routes) move to
server/api/inspections/hierarchy.ts. Handler bodies byte-identical; external
paths and OpenAPI/RPC surface unchanged.
